chore(e2e): Decoupled Currents api from Quarantine bot
