+++ packages/e2e-utils/src/currentsApi/config.ts
@@ -1,0 +1,3 @@
+export const CURRENTS_API_BASE = 'https://api.currents.dev/v1';
+export const TEST_RESULTS_PAGE_SIZE = 10; // default page size of the test-results API endpoint
+export const DEVELOP_BRANCH = 'develop';
