+++ packages/e2e-utils/src/quarantineBot/quarantine.ts
@@ -1,11 +1,6 @@
 /* eslint-disable no-console */
 import { computeStats, extractKeyFromAction, getTestKey, normalizeTitlePath } from './actions';
-import {
-    createQuarantineAction,
-    deleteAction,
-    getLastNResults,
-    getLastNResultsFromDistinctBranches,
-} from './api';
+import { createQuarantineAction } from './api';
 import {
     EXPLORER_LOOKBACK_DAYS,
     PRE_FILTER_FAILURE_RATE,
@@ -15,6 +10,7 @@ import {
     UNQUARANTINE_LAST_N_EXECUTIONS,
 } from './config';
 import type { Action, SlackEvent, TestExplorerItem } from './types';
+import { deleteAction, getLastNResults, getLastNResultsFromDistinctBranches } from '../currentsApi/api';
 
 export async function quarantineFailingTests(
     projectId: string,
@@ -52,7 +48,7 @@ export async function quarantineFailingTests(
             continue;
         }
 
-        const results = await getLastNResultsFromDistinctBranches(test.signature);
+        const results = await getLastNResultsFromDistinctBranches(test.signature, QUARANTINE_LAST_N_EXECUTIONS, EXPLORER_LOOKBACK_DAYS);
 
         if (results.length < QUARANTINE_LAST_N_EXECUTIONS) {
             console.log(
@@ -131,7 +127,7 @@ export async function unquarantinePassingTests(
             continue;
         }
 
-        const results = await getLastNResults(test.signature, UNQUARANTINE_LAST_N_EXECUTIONS);
+        const results = await getLastNResults(test.signature, UNQUARANTINE_LAST_N_EXECUTIONS, EXPLORER_LOOKBACK_DAYS);
 
         if (results.length < UNQUARANTINE_LAST_N_EXECUTIONS) {
             console.log(
